requirements.txt
# Web framework
fastapi[standard]
uvicorn[standard]

# WebSockets
websockets

# Data processing
numpy
pandas

# Machine Learning
scikit-learn
tensorflow
scapy

aio-pika
pymongo
python-dotenv